refactor: move environments files to environments library
