+++ apps/frontend/src/main.ts
@@ -2,7 +2,7 @@ import { enableProdMode, ViewEncapsulation } from '@angular/core'
 import { platformBrowserDynamic } from '@angular/platform-browser-dynamic'
 
 import { AppModule } from './app/app.module'
-import { environment } from './environments/environment'
+import { environment } from '../../../libs/root/environments/src/environments/environment'
 
 if (environment.production) {
   enableProdMode()
